modified to latest updates
