+++ main.py
@@ -1,47 +1,11 @@
-import numpy as np
 import pandas as pd
-import random
-import pycuda.driver as cuda
-from pycuda import driver, compiler, gpuarray
-import pycuda.autoinit
-import csv
-import matplotlib.pyplot as plt
 from sklearn.metrics import silhouette_score, davies_bouldin_score, calinski_harabasz_score
 from time import time
 
 from K_Means import k_means
-from GPU_K_Means import k_means_gpu
 from Dictionary_and_arrays_converts import (make_indicator_dictionary, make_capital_dictionary,
                                             convert_clusters_to_arrays, next_step_clustering_list)
-
-def plot_clusters(clusters, centroids, dataset_name, iteration=None):
-    colors = ['r', 'g', 'b', 'y', 'c', 'm', 'purple', 'orange']
-    markers = ['o', 's', 'D', '^', 'P', 'X', 'p']
-
-    # Plot clusters
-    for cluster_index, cluster_data in clusters.items():
-        cluster_points = [list(item.values())[0] for item in cluster_data]
-        cluster_points = np.array(cluster_points)
-        plt.scatter(cluster_points[:, 0], cluster_points[:, 1],
-                    c=colors[cluster_index % len(colors)],
-                    marker=markers[cluster_index % len(markers)],
-                    label=f'Cluster {cluster_index}')
-
-    # Plot centroids
-    for idx, centroid in enumerate(centroids):
-        centroid_values = list(centroid.values())[0]
-        plt.scatter(centroid_values[0], centroid_values[1],
-                    c='black', marker='x', s=100, linewidths=3,
-                    label=f'Centroid {idx}')
-
-    if iteration == 0:
-        plt.title(f'{dataset_name} Clusters (First iteration)')
-    else:
-        plt.title(f'{dataset_name} Clusters')
-
-    plt.legend(bbox_to_anchor=(1.05, 1), loc='upper left')
-    plt.tight_layout()
-    plt.show()
+from plot import plot_clusters
 
 
 
@@ -51,7 +15,7 @@ def show_clusters(clusters):
         print(f"Cluster {cluster_index}: {country_names}")
 
 
-def evaluate_clustering(X, labels):
+def output_scores(X, labels):
     # Оцінюємо за допомогою різних метрик:
     silhouette_avg = silhouette_score(X, labels)
     print(f"Силуетний індекс: {silhouette_avg}")
@@ -60,52 +24,15 @@ def evaluate_clustering(X, labels):
     print(f"Індекс Девіса-Болдіна: {db_score}")
 
     ch_score = calinski_harabasz_score(X, labels)
-    print(f"Індекс Каліна-Харабаша: {ch_score}")
+    print(f"Індекс Каліна-Харабаша: {ch_score}\n")
 
-
-
+k = 7
 
 if __name__ == "__main__":
+
+    # --- --- -- Capitals --- --- ---
+
     # Читаємо CSV-файл
-    df = pd.read_csv('Worldbank-data-all.csv')
-
-    first_column = df.columns[0]  # Перша колонка містить назву країни
-    last_column = df.columns[-1]  # Остання колонка містить числові дані
-    grouped_with_country = make_indicator_dictionary(first_column, last_column)
-    print(f"Countries quantity(indicators): {len(grouped_with_country)}")
-    k = 7  # Кількість кластерів
-
-    # Час виконання на CPU
-    start_time = time()
-    clusters_i, centroids_i = k_means(grouped_with_country, k)
-    cpu_time = time() - start_time
-
-    print("Clusters (CPU):")
-    show_clusters(clusters_i)
-    print(f"CPU Time: {cpu_time:.5f} seconds")
-    print()
-    plot_clusters(clusters_i, centroids_i, 'Capitals')  # Plot відображення утворених кластерів
-
-    # вивід метрик
-    X_i, labels_i = convert_clusters_to_arrays(clusters_i)
-    evaluate_clustering(X_i, labels_i)
-
-    start_time = time()
-    clusters_ig, centroids_ig = k_means_gpu(grouped_with_country, k)
-    elapsed_time = time() - start_time
-
-    print("Clusters(GPU):")
-    show_clusters(clusters_ig)
-    print(f"GPU Time: {elapsed_time:.5f} seconds")
-    print()
-
-    # вивід метрик
-    X_ig, labels_ig = convert_clusters_to_arrays(clusters_ig)
-    print("GPU Evaluation:")
-    evaluate_clustering(X_ig, labels_ig)
-
-    # --- ---- Столиці ---- ---
-
     capitals = pd.read_csv('capitals-location.csv')
     name_column = capitals.columns[0]  # Перша колонка містить назву країни
     coordinate_column = capitals.columns[2:4]  # Остання колонка містить числові дані
@@ -113,72 +40,92 @@ if __name__ == "__main__":
     grouped_with_country = make_capital_dictionary(name_column, coordinate_column)
     print("\n --- --- --- --- --- --- --- --- --- --- ---\n")
     print(f"Countries quantity(capitals): {len(grouped_with_country)}")
-    k = 7  # Кількість кластерів
 
     # Час виконання на CPU
-    start_time = time()
-    clusters_f, centroids_f = k_means(grouped_with_country, k)
-    cpu_time = time() - start_time
+    start_time_f = time()
+    clusters_f, centroids_f = k_means(grouped_with_country, k, 'CPU')
+    cpu_time_f = time() - start_time_f
 
-    print("Clusters (CPU):")
+    print("Capitals clusters (CPU):")
     show_clusters(clusters_f)
-    print(f"CPU Time: {cpu_time:.5f} seconds")
+    print(f"CPU Time: {cpu_time_f:.5f} seconds")
     print()
-    plot_clusters(clusters_f, centroids_f, 'Capitals')  # Plot відображення утворених кластерів
+    plot_clusters(clusters_f, centroids_f, 'Capitals')
+
+    # вивід метрик
+    X_f, labels_f = convert_clusters_to_arrays(clusters_f)
+    output_scores(X_f, labels_f)
 
     # Час виконання на GPU
-    start_time = time()
-    clusters_s, centroids_s = k_means_gpu(grouped_with_country, k)
-    elapsed_time = time() - start_time
+    start_time_f_gpu = time()
+    clusters_f_gpu, centroids_f_gpu = k_means(grouped_with_country, k, 'GPU')
+    gpu_time_f = time() - start_time_f_gpu
 
-    print("Clusters(GPU):")
+    print("Capitals clusters (GPU):")
+    show_clusters(clusters_f_gpu)
+    print(f"GPU Time: {gpu_time_f:.5f} seconds")
+    print()
+    plot_clusters(clusters_f_gpu, centroids_f_gpu, 'Capitals')
+
+    # вивід метрик
+    X_f_gpu, labels_f_gpu = convert_clusters_to_arrays(clusters_f_gpu)
+    output_scores(X_f_gpu, labels_f_gpu)
+
+    # --- --- -- Indicators --- --- ---
+
+    df = pd.read_csv('Worldbank-data-all.csv')
+    first_column = df.columns[0]  # Перша колонка містить назву країни
+    last_column = df.columns[-1]  # Остання колонка містить числові дані
+    grouped_with_country = make_indicator_dictionary(first_column, last_column)
+    print(f"Countries quantity(indicators): {len(grouped_with_country)}")
+
+    start_time_s = time()
+    clusters_s, centroids_s = k_means(grouped_with_country, k, 'CPU')
+    cpu_time = time() - start_time_s
     show_clusters(clusters_s)
+    print(f"CPU Time: {cpu_time:.5f} seconds")
+    print()
+
+    X_s, labels_s = convert_clusters_to_arrays(clusters_s)
+    output_scores(X_s, labels_s)
+
+    start_time_s_gpu = time()
+    clusters_s_gpu, centroids_s_gpu = k_means(grouped_with_country, k, 'GPU')
+    elapsed_time = time() - start_time_s_gpu
+    show_clusters(clusters_s_gpu)
     print(f"GPU Time: {elapsed_time:.5f} seconds")
     print()
-    #plot_clusters(clusters_s, centroids_s, 'Capitals')  # Plot відображення утворених кластерів
 
-    # Метрики для обох підходів
-    X_gpu, labels_gpu = convert_clusters_to_arrays(clusters_s)
-    print("GPU Evaluation:")
-    evaluate_clustering(X_gpu, labels_gpu)
+    X_s_gpu, labels_s_gpu = convert_clusters_to_arrays(clusters_s_gpu)
+    output_scores(X_s_gpu, labels_s_gpu)
 
+    # --- --- -- Combine clusterization --- --- ---
 
-    # --- ---- Комбінована кластеризація ---- ---
-
-    nex_step_dict = next_step_clustering_list(clusters_i, clusters_f)
+    nex_step_dict = next_step_clustering_list(clusters_f, clusters_s)
     print("\n --- --- --- --- --- --- --- --- --- --- ---\n")
     print(f"Indicators and capitals(Combine): {len(grouped_with_country)}")
-    k = 7  # Кількість кластерів
 
-
-    # Час виконання на CPU
-    start_time = time()
-    clusters_l, centroids_l = k_means(nex_step_dict, k)
-    cpu_time = time() - start_time
-
-    print("Clusters (CPU):")
+    start_time_l = time()
+    clusters_l, centroids_l = k_means(nex_step_dict, k, 'CPU')
+    cpu_time = time() - start_time_l
     show_clusters(clusters_l)
     print(f"CPU Time: {cpu_time:.5f} seconds")
     print()
-    plot_clusters(clusters_l, centroids_l, 'Capitals')  # Plot відображення утворених кластерів
 
-    X_gpu, labels_gpu = convert_clusters_to_arrays(clusters_l)
-    print("GPU Evaluation:")
-    evaluate_clustering(X_gpu, labels_gpu)
+    X_l, labels_l = convert_clusters_to_arrays(clusters_l)
+    output_scores(X_l, labels_l)
+
+    print("--- --- --- Розпаралелений обрахунок --- --- ---")
     print()
 
-    print("--- --- --- Розпаралелений результат --- --- ---")
-    # Час виконання на GPU
-    start_time = time()
-    clusters_lg, centroids_lg = k_means_gpu(nex_step_dict, k)
-    elapsed_time = time() - start_time
-
-    print("Clusters(GPU):")
-    show_clusters(clusters_lg)
+    start_time_l_gpu = time()
+    clusters_l_gpu, centroids_l_gpu = k_means(nex_step_dict, k, 'GPU')
+    elapsed_time = time() - start_time_l_gpu
+    show_clusters(clusters_l_gpu)
     print(f"GPU Time: {elapsed_time:.5f} seconds")
     print()
 
-    # Метрики
-    X_gpu, labels_gpu = convert_clusters_to_arrays(clusters_lg)
-    print("GPU Evaluation:")
-    evaluate_clustering(X_gpu, labels_gpu)
+    X_l_gpu, labels_l_gpu = convert_clusters_to_arrays(clusters_l_gpu)
+    output_scores(X_l_gpu, labels_l_gpu)
+
+    plot_clusters(clusters_l_gpu, centroids_l_gpu, 'Combine')
